refactor: move file to source root
